src/app/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
databricks-sql-connector>=3.1.0
